Create 43. Multiply Strings
This is a medium question. When I have the question, I write on the paper to figure out the natural calculation steps. With the practice of plus method, I realize how to write the multiple. So the divide and conquer methodology also help me to slice the big question into small pieces. The code is the realization of your thinking process. But many hidden part will not be seen just like the iceberg.