Create 190. Reverse Bits
I need to remember the bitwise & 1 to get the % operation!

>>> means add 0 on the left
>>  means add 1 on the left
">>" is signed because it keeps the sign. It uses the most left digit in binary representation of a number as a filler. For example:
      | this value is used as a filler 
      11011011 
 >> 11101101  

      01010010
 >> 00101001 
">>>" is unsigned version of this operator. It always use zero as a filler:

       11011011 
>>> 01101101  

       01010010
>>> 00101001

In binary representation the most left digit determines sign of the number. So, if it's '1' then we have negative value and if it's '0' - then our number is positive. That's why using the most left digit as a filler allows to keep sign permanent.