Create 310. Minimum Height Trees
This is a typical graph problem. You need to understand the algorithm first. Then you need to use the collections class to implement the algorithm. It is much better to draw the graph on paper to give you a good plan to write the program.

it is very similar to the course schedule question.